backend/RAG_chatbot/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
starlette==0.36.3

# Database
sqlalchemy==2.0.27
pymysql==1.1.1
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment and config
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.6.0

# RAG and LLM - Phi-3/DeepSeek with Ollama
langchain==0.3.0
langchain-community==0.3.0
ollama==0.6.0

# Vector Database and Embeddings
chromadb==0.5.18
sentence-transformers==3.3.1
faiss-cpu==1.9.0
posthog==3.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
pypdf==5.1.0
pdfplumber==0.11.4

# Excel Processing
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-calamine==0.2.3

# Google Drive Integration
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Web Scraping and URL Processing
beautifulsoup4==4.12.3
html2text==2024.2.26
requests==2.32.3
gdown==5.2.0
lxml[html_clean]==5.3.0
selenium==4.27.1
trafilatura==1.12.2
webdriver-manager==4.0.2

# Image Processing and OCR
Pillow==11.0.0
pytesseract==0.3.13
pdf2image==1.17.0
opencv-python==4.10.0.84

# File handling
aiofiles==24.1.0
httpx==0.27.2

# Text processing
nltk==3.9.1
spacy==3.8.2
tiktoken==0.8.0

# Core Data Science - FIXED VERSIONS
numpy==1.26.4
scipy==1.14.1

# Visualization - MATPLOTLIB & SEABORN
matplotlib==3.9.3
seaborn==0.13.2

# Utilities
huggingface-hub==0.26.5
email-validator==2.2.0
validators==0.34.0

# Additional dependencies
contourpy==1.3.1
pyodbc==5.0.1